+++ pkgs/multistockfish/lib/src/stockfish_flavor.dart
@@ -1,7 +1,10 @@
 /// The flavor of Stockfish to use.
 enum StockfishFlavor {
-  /// Regular NNUE Stockfish engine for chess and chess960
-  chess,
+  /// Stockfish engine version 16 with embedded NNUE
+  sf16,
+
+  /// Latest Stockfish engine version without embedded NNUE
+  latestNoNNUE,
 
   /// Multi-Variant Stockfish using Handcrafted Evaluation for chess and chess variants
   variant,