+++ pkgs/multistockfish_sf16/ios/Classes/stockfish16.h
@@ -1,0 +1,36 @@
+#include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+#if _WIN32
+#include <windows.h>
+#else
+#include <pthread.h>
+#include <unistd.h>
+#endif
+
+#if _WIN32
+#define FFI_PLUGIN_EXPORT __declspec(dllexport)
+#else
+#define FFI_PLUGIN_EXPORT
+#endif
+
+#ifdef __cplusplus
+extern "C" __attribute__((visibility("default"))) __attribute__((used))
+#endif
+FFI_PLUGIN_EXPORT int stockfish_init();
+
+#ifdef __cplusplus
+extern "C" __attribute__((visibility("default"))) __attribute__((used))
+#endif
+FFI_PLUGIN_EXPORT int stockfish_main();
+
+#ifdef __cplusplus
+extern "C" __attribute__((visibility("default"))) __attribute__((used))
+#endif
+FFI_PLUGIN_EXPORT ssize_t stockfish_stdin_write(char *data);
+
+#ifdef __cplusplus
+extern "C" __attribute__((visibility("default"))) __attribute__((used))
+#endif
+FFI_PLUGIN_EXPORT char * stockfish_stdout_read();